Change player creation logic to resolve dependancy loops
